Fix test, move spec to WorkspaceWarningIcon.test
